requirements.txt
fastapi
uvicorn
openai-whisper
langdetect
yt_dlp
google-genai

